make loadPlugin a method of boot decouple plugin from boot
