+++ lib/plugin.js
@@ -1,24 +1,22 @@
 'use strict'
 
-const fastq = require('fastq')
 const { EventEmitter } = require('events')
 const { inherits } = require('util')
 const { debug } = require('./debug')
-const { loadPlugin } = require('./load-plugin')
 const { createPromise } = require('./create-promise')
 const { AVV_ERR_READY_TIMEOUT } = require('./errors')
 const { getPluginName } = require('./get-plugin-name')
 const { isPromiseLike } = require('./is-promise-like')
 
 /**
- * @param {*} parent
+ * @param {fastq.queue} queue
  * @param {*} func
  * @param {*} options
  * @param {boolean} isAfter
  * @param {number} [timeout]
  */
-function Plugin (parent, func, options, isAfter, timeout) {
-  this.parent = parent
+function Plugin (queue, func, options, isAfter, timeout) {
+  this.queue = queue
   this.func = func
   this.options = options
   /**
@@ -39,10 +37,6 @@ function Plugin (parent, func, options, isAfter, timeout) {
    */
   this.name = getPluginName(func, options)
 
-  /**
-   * @type {fastq.queue}
-   */
-  this.queue = fastq(parent, loadPluginNextTick, 1)
   this.queue.pause()
 
   /**
@@ -263,14 +257,6 @@ Plugin.prototype.finish = function (err, callback) {
   // we start loading the dependents plugins only once
   // the current level is finished
   this.queue.resume()
-}
-
-/**
- * Delays plugin loading until the next tick to ensure any bound `_after` callbacks have a chance
- * to run prior to executing the next plugin
- */
-function loadPluginNextTick (plugin, callback) {
-  process.nextTick(loadPlugin, this, plugin, callback)
 }
 
 function noop () {}
