+++ test/load-plugin.test.js
@@ -1,19 +1,19 @@
 'use strict'
 
+const fastq = require('fastq')
+const boot = require('..')
 const { test } = require('tap')
-const boot = require('../..')
-const { loadPlugin } = require('../../lib/load-plugin')
-const { Plugin } = require('../../lib/plugin')
+const { Plugin } = require('../lib/plugin')
 
 test('successfully load a plugin with sync function', (t) => {
   t.plan(1)
   const app = boot({})
 
-  const plugin = new Plugin(app, function (instance, opts, done) {
+  const plugin = new Plugin(fastq(app, app.loadPluginNextTick, 1), function (instance, opts, done) {
     done()
   }, false, 0)
 
-  loadPlugin(app, plugin, function (err) {
+  app.loadPlugin(plugin, function (err) {
     t.equal(err, undefined)
   })
 })
@@ -22,11 +22,11 @@ test('catch an error when loading a plugin with sync function', (t) => {
   t.plan(1)
   const app = boot({})
 
-  const plugin = new Plugin(app, function (instance, opts, done) {
+  const plugin = new Plugin(fastq(app, app.loadPluginNextTick, 1), function (instance, opts, done) {
     done(Error('ArbitraryError'))
   }, false, 0)
 
-  loadPlugin(app, plugin, function (err) {
+  app.loadPlugin(plugin, function (err) {
     t.equal(err.message, 'ArbitraryError')
   })
 })
@@ -35,9 +35,9 @@ test('successfully load a plugin with async function', (t) => {
   t.plan(1)
   const app = boot({})
 
-  const plugin = new Plugin(app, async function (instance, opts) { }, false, 0)
+  const plugin = new Plugin(fastq(app, app.loadPluginNextTick, 1), async function (instance, opts) { }, false, 0)
 
-  loadPlugin(app, plugin, function (err) {
+  app.loadPlugin(plugin, function (err) {
     t.equal(err, undefined)
   })
 })
@@ -46,11 +46,11 @@ test('catch an error when loading a plugin with async function', (t) => {
   t.plan(1)
   const app = boot({})
 
-  const plugin = new Plugin(app, async function (instance, opts) {
+  const plugin = new Plugin(fastq(app, app.loadPluginNextTick, 1), async function (instance, opts) {
     throw Error('ArbitraryError')
   }, false, 0)
 
-  loadPlugin(app, plugin, function (err) {
+  app.loadPlugin(plugin, function (err) {
     t.equal(err.message, 'ArbitraryError')
   })
 })
@@ -59,11 +59,11 @@ test('successfully load a plugin when function is a Promise, which resolves to a
   t.plan(1)
   const app = boot({})
 
-  const plugin = new Plugin(app, new Promise(resolve => resolve(function (instance, opts, done) {
+  const plugin = new Plugin(fastq(app, app.loadPluginNextTick, 1), new Promise(resolve => resolve(function (instance, opts, done) {
     done()
   })), false, 0)
 
-  loadPlugin(app, plugin, function (err) {
+  app.loadPlugin(plugin, function (err) {
     t.equal(err, undefined)
   })
 })
@@ -72,11 +72,11 @@ test('catch an error when loading a plugin when function is a Promise, which res
   t.plan(1)
   const app = boot({})
 
-  const plugin = new Plugin(app, new Promise(resolve => resolve(function (instance, opts, done) {
+  const plugin = new Plugin(fastq(app, app.loadPluginNextTick, 1), new Promise(resolve => resolve(function (instance, opts, done) {
     done(Error('ArbitraryError'))
   })), false, 0)
 
-  loadPlugin(app, plugin, function (err) {
+  app.loadPlugin(plugin, function (err) {
     t.equal(err.message, 'ArbitraryError')
   })
 })
@@ -85,13 +85,13 @@ test('successfully load a plugin when function is a Promise, which resolves to a
   t.plan(1)
   const app = boot({})
 
-  const plugin = new Plugin(app, new Promise(resolve => resolve({
+  const plugin = new Plugin(fastq(app, app.loadPluginNextTick, 1), new Promise(resolve => resolve({
     default: function (instance, opts, done) {
       done()
     }
   })), false, 0)
 
-  loadPlugin(app, plugin, function (err) {
+  app.loadPlugin(plugin, function (err) {
     t.equal(err, undefined)
   })
 })
@@ -100,13 +100,13 @@ test('catch an error when loading a plugin when function is a Promise, which res
   t.plan(1)
   const app = boot({})
 
-  const plugin = new Plugin(app, new Promise(resolve => resolve({
+  const plugin = new Plugin(fastq(app, app.loadPluginNextTick, 1), new Promise(resolve => resolve({
     default: function (instance, opts, done) {
       done(Error('ArbitraryError'))
     }
   })), false, 0)
 
-  loadPlugin(app, plugin, function (err) {
+  app.loadPlugin(plugin, function (err) {
     t.equal(err.message, 'ArbitraryError')
   })
 })
